+++ src/LoginPage/LoginPage.jsx
@@ -1,11 +1,7 @@
-import React, { useState } from 'react';
-import { makeStyles } from "@material-ui/styles";
-import PropTypes from 'prop-types';
-import classNames from 'classnames';
-import { withStyles } from '@material-ui/core/styles';
-import MenuItem from '@material-ui/core/MenuItem';
-import TextField from '@material-ui/core/TextField';
 import Button from '@material-ui/core/Button';
+import TextField from '@material-ui/core/TextField';
+import { makeStyles } from "@material-ui/styles";
+import React, { useState } from 'react';
 
 
 const useStyles = makeStyles(theme => ({
