+++ src/RegisterUserPage/RegisterUserPage.jsx
@@ -1,11 +1,5 @@
 import React from 'react';
-import { makeStyles } from "@material-ui/styles";
-
-const useStyles = makeStyles(theme => ({}));
 
 export const RegisterUserPage = () => {
-
-
-    const classes = useStyles();
     return (<div>RegisterUserPage</div>);
 }
